+++ database/migrations/24-v_deviation_cykelparkeringsplatser_gavle.sql
@@ -27,8 +27,8 @@ ALTER TABLE upstream.mv_match_cykelparkeringsplatser_gavle OWNER TO app;
 
 CREATE OR REPLACE VIEW upstream.v_deviation_cykelparkeringsplatser_gavle AS
 	SELECT
-		21 AS dataset_id,
-		16 AS layer_id,
+		21::bigint AS dataset_id,
+		16::bigint AS layer_id,
 		upstream_item_ids,
 		CASE
 			WHEN osm_element_id IS NULL THEN upstream_geom
