+++ database/migrations/25-v_deviation_historiskaskyltar_gavle.sql
@@ -30,8 +30,8 @@ ALTER TABLE upstream.mv_match_historiskaskyltar_gavle OWNER TO app;
 CREATE OR REPLACE VIEW upstream.v_deviation_historiskaskyltar_gavle AS
 	SELECT *
     FROM (SELECT DISTINCT ON (match_id)
-		27 AS dataset_id,
-		16 AS layer_id,
+		27::bigint AS dataset_id,
+		16::bigint AS layer_id,
 		upstream_item_ids,
 		CASE
 			WHEN osm_element_id IS NULL THEN upstream_geom
