+++ database/migrations/27-v_deviation_byggnadsanlaggningspunkt_topo50.sql
@@ -25,8 +25,8 @@ ALTER TABLE upstream.mv_match_byggnadsanlaggningspunkt_topo50 OWNER TO app;
 DROP VIEW IF EXISTS upstream.v_deviation_byggnadsanlaggningspunkt_topo50;
 CREATE OR REPLACE VIEW upstream.v_deviation_byggnadsanlaggningspunkt_topo50 AS
 	SELECT
-		143 AS dataset_id,
-		19 AS layer_id,
+		143::bigint AS dataset_id,
+		19::bigint AS layer_id,
 		upstream_item_ids,
 		CASE
 			WHEN osm_element_id IS NULL THEN upstream_geom
