+++ database/migrations/35-v_deviation_busshallplatser_gavle.sql
@@ -41,8 +41,8 @@ ALTER TABLE upstream.mv_match_busshallplatser_gavle OWNER TO app;
 
 CREATE OR REPLACE VIEW upstream.v_deviation_busshallplatser_gavle AS
 	SELECT
-		18 AS dataset_id,
-		17 AS layer_id,
+		18::bigint AS dataset_id,
+		17::bigint AS layer_id,
 		upstream_item_ids,
 		CASE
 			WHEN osm_element_id IS NULL THEN upstream_geom
